Basket에 amount 모델 변경 PositiveSmallIntegerField

+++ app/bill/models.py
@@ -21,7 +21,7 @@ class Basket(models.Model):
         null=True,
         blank=True,
     )
-    amount = models.IntegerField(default=0)
+    amount = models.PositiveSmallIntegerField(default=0)
     order_yn = models.BooleanField(default=False)
 
     def __str__(self):
